Add an ebuild for onedriver.
Added an ebuild for onedriver, a Linux program made by jstaf that allows to access OneDrive files from the file manager. Right now it works just fine whether the "gui" flag is used or not, but it can probably be improved, specially since the hard dependency on systemd isn't inmediately clear and I should maybe add it as a depedency on the ebuild.

This release includes a patch to disable tests on the Makefile since they require to be run with sudo, which is not allowed by Portage. It comes after I uploaded a dependency tarball for onedriver a few days back, since onedriver does not include as of v0.13.0 a vendor directory and makes it a necessary step to build it.